requirements.txt
matplotlib==3.7.5
numpy==1.24.3
scikit-learn==1.3.2
thop==0.1.1.post2209072238
torchaudio==0.13.1+cu116
tqdm==4.66.4
